rename: Location Icon 폴더명 변경
